refactored the Integrate files. added Formik to vvalidate forms. added new designs and styled
