--- components/Layout.js
@@ -1,14 +1,0 @@
-import { useState } from 'react';
-import Head from "next/head";
-import { useRouter } from "next/router";
-import Link from 'next/link';
-import Blob from "../components/Blob";
-
-const Layout = () => {
-    return (
-        <>
-        </>
-    )
-}
-
-export default Layout;
