+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@textile/hub": "^6.2.0",
     "antd": "^4.16.6",
+    "dotenv": "^10.0.0",
     "ethers": "^5.4.1",
     "formik": "^2.2.9",
     "formik-antd": "^2.0.1",
